+++ ext/oj/mimic_json.c
@@ -211,6 +211,7 @@ static VALUE mimic_dump(int argc, VALUE *argv, VALUE self) {
     out.buf           = buf;
     out.end           = buf + sizeof(buf) - 10;
     out.allocated     = false;
+    out.generate      = false;
     out.caller        = CALLER_DUMP;
     copts.escape_mode = JXEsc;
     copts.mode        = CompatMode;
@@ -371,6 +372,7 @@ static VALUE mimic_generate_core(int argc, VALUE *argv, Options copts) {
     out.allocated = false;
     out.omit_nil  = copts->dump_opts.omit_nil;
     out.caller    = CALLER_GENERATE;
+    out.generate  = true;
     // For obj.to_json or generate nan is not allowed but if called from dump
     // it is.
     copts->dump_opts.nan_dump = RaiseNan;
@@ -746,6 +748,7 @@ static VALUE mimic_object_to_json(int argc, VALUE *argv, VALUE self) {
     out.end           = buf + sizeof(buf) - 10;
     out.allocated     = false;
     out.omit_nil      = copts.dump_opts.omit_nil;
+    out.generate      = false;
     copts.mode        = CompatMode;
     copts.to_json     = No;
     if (1 <= argc && Qnil != argv[0]) {
@@ -755,7 +758,7 @@ static VALUE mimic_object_to_json(int argc, VALUE *argv, VALUE self) {
     // seem to prefer the option of changing that.
     // oj_dump_obj_to_json(self, &mimic_object_to_json_options, &out);
     oj_dump_obj_to_json_using_params(self, &copts, &out, argc, argv);
-    if (0 == out.buf) {
+    if (NULL == out.buf) {
         rb_raise(rb_eNoMemError, "Not enough memory.");
     }
     rstr = rb_str_new2(out.buf);